+++ packages/sources/src/components/image-overlay/image-overlay.args.ts
@@ -1,0 +1,10 @@
+import { ArgTypes } from "../../stories-helpers";
+import { ImageOverlay } from "./image-overlay.models";
+
+export interface ImageOverlayArgs {}
+
+export const imageOverlayArgTypes: ArgTypes<ImageOverlayArgs> = {};
+
+export function imageOverlayArgsMapper(a: ImageOverlayArgs): ImageOverlay {
+  return {};
+}